Solution PartA-2

+++ src/main/java/com/upgrad/imagehoster/controller/ImageController.java
@@ -105,13 +105,20 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imageOwner = image.getUser();
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+       if(currentUser.getId() == imageOwner.getId()){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }else{
+           model.addAttribute("editError", true);
+            return "redirect:/images/" + imageId;
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -145,7 +152,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId();
     }
 
 
@@ -153,9 +160,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session) {
+
+
+        Image image = imageService.getImage(imageId);
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imageOwner = image.getUser();
+
+        if(currentUser.getId() == imageOwner.getId()){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }else{
+            //model.addAttribute("deleteError", true);
+            return "redirect:/images/" + imageId;
+        }
     }
